Update Ws28xx project
