+++ src/devices/Ws28xx/samples/Ws28xx.Samples.csproj
@@ -2,7 +2,7 @@
 
    <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
